Actions on index page

+++ app/models/key.rb
@@ -3,8 +3,8 @@ require 'rufus-scheduler'
 
 class Key < ActiveRecord::Base
   before_save :generate_key
-  @ttl = 30;
-  @ttb = 5;
+  @ttl = 300;
+  @ttb = 50;
 
   def generate_key
     key = SecureRandom.urlsafe_base64
@@ -14,15 +14,16 @@ class Key < ActiveRecord::Base
   end
 
   scheduled_cleanup = Rufus::Scheduler.new
-  scheduled_cleanup.every '1s' do
+  scheduled_cleanup.every '10s' do
     keys = Key.all
-    keys.each do |key|
-      puts key.value
-      curr_time = Time.now.to_i
-      if(curr_time - key.alive_time > @ttl)
-        key.destroy
-      elsif(key.block_time !=0 && curr_time - key.block_time > @ttb)
-        key.update_column(:block_time, 0)
+    if(!keys.empty?)
+      keys.each do |key|
+        curr_time = Time.now.to_i
+        if(curr_time - key.alive_time > @ttl)
+          key.destroy
+        elsif(key.block_time !=0 && curr_time - key.block_time > @ttb)
+          key.update_column(:block_time, 0)
+        end
       end
     end
   end
